had some merge issues. i think we're back to normal. users, roles, and sessions have session verify checks now
